refactor: rename instruction-cotainer.tsx to instruction-container.tsx for consistency with component name
